Add files via upload
This program is a PNG encoder that takes a PNG file, input.png, and produces a file, compressed.png, as output.  Update the string, sourceImagePath, with the appropriate location for your input image.

Updates to this program that I am working on include:
*  The implementation of the parallel processing of image blocks to speed up the encoder.
*  The creation of a windows graphical user interface for the program.

input.png image credit: pixel2013
https://pixabay.com/images/id-4777410/

Simplified Pixabay License
Our license empowers creators and protects our community. We want to keep it as simple as possible. Here is an overview of what Pixabay content can and can't be used for.
What is allowed?
*  All content on Pixabay can be used for free for commercial and noncommercial use across print and digital, except in the cases mentioned in "What is not allowed".
*  Attribution is not required. Giving credit to the contributor or Pixabay is not necessary but is always appreciated by our community.
*  You can make modifications to content from Pixabay.